requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.8
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
